rename mainpage folder to camelcase
