+++ src/components/mainPage/customCarousel.js
